+++ tree-sitter-stack-graphs/tests/it/builder.rs
@@ -27,10 +27,10 @@ fn can_support_preexisting_nodes() {
     let node_id = graph.new_node_id(file);
     let _preexisting_node = graph.add_scope_node(node_id, true).unwrap();
 
-    let mut globals = Variables::new();
+    let globals = Variables::new();
     let language = StackGraphLanguage::from_str(tree_sitter_python::language(), tsg).unwrap();
     language
-        .build_stack_graph_into(&mut graph, file, python, &mut globals, &NoCancellation)
+        .build_stack_graph_into(&mut graph, file, python, &globals, &NoCancellation)
         .expect("Failed to build graph");
 }
 
@@ -59,7 +59,7 @@ fn can_support_injected_nodes() {
         .expect("Failed to add EXT_NODE variable");
 
     builder
-        .build(&mut globals, &NoCancellation)
+        .build(&globals, &NoCancellation)
         .expect("Failed to build graph");
 
     check_stack_graph_nodes(