+++ tree-sitter-stack-graphs/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Library
+
+#### Fixed
+
+- Executing a `StackGraphLanguage` for a `File` for which the `StackGraph` already contains nodes does not crash anymore.
+
+#### Added
+
+- The `StackGraphLanguage::builder_into_stack_graph` method can be used to create a `Builer` that allows injecting preexisting `StackGraph` nodes using `Builder::inject_node` to obtain `tree_sitter_graph::graph::Value` instances. These can be used for global variables such that the TSG rules can refer to stack graph nodes that were not created by them.
+
 ## 0.3.0 -- 2022-08-23
 
 ### Library