+++ tree-sitter-stack-graphs/CHANGELOG.md
@@ -11,11 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Fixed
 
-- Executing a `StackGraphLanguage` for a `File` for which the `StackGraph` already contains nodes does not crash anymore.
+- Cancellation errors from `tree_sitter_graph` execution are not wrapped
+  but propagated as cancellations.
+- Executing a `StackGraphLanguage` for a `File` for which the `StackGraph`
+  already contains nodes does not crash anymore.
 
 #### Added
 
-- The `StackGraphLanguage::builder_into_stack_graph` method can be used to create a `Builer` that allows injecting preexisting `StackGraph` nodes using `Builder::inject_node` to obtain `tree_sitter_graph::graph::Value` instances. These can be used for global variables such that the TSG rules can refer to stack graph nodes that were not created by them.
+- The `StackGraphLanguage::builder_into_stack_graph` method can be used to
+  create a `Builer` that allows injecting preexisting `StackGraph` nodes
+  using `Builder::inject_node` to obtain `tree_sitter_graph::graph::Value`
+  instances. These can be used for global variables such that the TSG
+  rules can refer to stack graph nodes that were not created by them.
 
 #### Changed